--- a/app/helpers/datasets_helper.rb
+++ b/app/helpers/datasets_helper.rb
@@ -11,12 +11,6 @@ module DatasetsHelper
     dataset.distributions.reject { |distribution| distribution.refining? || distribution.broke? }
   end
 
-  def next_dataset(dataset)
-    datasets = current_organization.catalog.catalog_datasets.sort_by(&:publish_date)
-    index = datasets.index { |ds| ds.id == dataset.id }
-    datasets[index + 1]
-  end
-
   def sector_title(dataset)
     dataset.sector.present? ? dataset.sector.title : 'Sector no definido'
   end