--- a/app/views/datasets/edit.html.haml
+++ b/app/views/datasets/edit.html.haml
@@ -101,11 +101,6 @@
 .row
   .col-md-9.margin-top-20
     = link_to 'Volver al catálogo', catalog_datasets_path(current_organization.catalog), class: 'btn btn-alt btn-default btn-lg'
-    - if (next_dataset = next_dataset(@dataset))
-      = link_to edit_dataset_path(next_dataset), { class: 'btn btn-alt btn-default btn-lg pull-right' } do
-        Siguiente conjunto
-        %span.icon.icon-left
-          = inline_svg 'next.svg'
     .clearfix
     %br