--- a/app/views/distributions/edit.html.haml
+++ b/app/views/distributions/edit.html.haml
@@ -14,7 +14,7 @@
         = f.label :description, class: 'control-label'
         = f.text_area :description, required: true, class: 'autosize form-control', 'data-toggle': 'tooltip', title: "#{t('tooltip.dataset.description')}"
       .form-group.required
-        = f.label :download_url, 'URL para descargar'
+        = f.label :download_url, 'URL para descargar', class: 'control-label'
         = f.url_field :download_url, class: 'form-control', required: true, 'data-toggle': 'tooltip', title: "#{t('tooltip.distribution.download_url')}"
       .row
         .col-xs-6