--- a/app/views/inventories/datasets/index.html.haml
+++ b/app/views/inventories/datasets/index.html.haml
@@ -25,7 +25,30 @@
           %th.center= sortable('publish_date', 'Fecha estimada de publicación')
           %th.center Acciones
       %tbody
-        = render partial: 'inventories/shared/dataset', collection: @datasets
+        - @datasets.each_with_index do |dataset, index|
+          %tr.active.dataset
+            %td.center
+              %strong= index + 1
+            %td
+              %a.accordion-toggle{'data-toggle' => 'collapse', 'href' => ".dataset_#{dataset.id}_distributions"}
+                = dataset.title
+                %span.caret
+            %td.center= pluralize(dataset.distributions.count, 'recurso')
+            %td.center= dataset.public_access? ? 'Público' : 'Privado'
+            %td.center= dataset.publish_date.strftime('%F') if dataset.publish_date
+            %td.col-md-3
+              = render partial: 'inventories/shared/menus/dataset_actions', locals: { dataset: dataset }
+          = render partial: 'inventories/shared/distribution', collection: dataset.distributions.includes(:dataset)
+          %tr.accordion-body.collapse{'class' => "dataset_#{dataset.id}_distributions"}
+            %td
+            %td
+            %td
+            %td
+            %td
+            %td.add-distribution
+              = link_to new_inventories_dataset_distribution_path(dataset), { class: 'btn btn-primary btn-block' } do
+                %span.icon.glyphicon.glyphicon-plus{'aria-hidden': 'true'}
+                Agregar un recurso nuevo
   .row.margin-top-20
     .col-md-12
       = link_to organization_inventory_path(current_organization, format: :csv), { class: 'btn btn-primary btn-default btn-lg' } do