--- a/app/views/inventories/datasets/index.html.haml
+++ b/app/views/inventories/datasets/index.html.haml
@@ -33,7 +33,7 @@
               %a.accordion-toggle{'data-toggle' => 'collapse', 'href' => ".dataset_#{dataset.id}_distributions"}
                 = dataset.title
                 %span.caret
-            %td.center= pluralize(dataset.distributions.count, 'recurso')
+            %td.center= pluralize(dataset.distributions.length, 'recurso')
             %td.center= dataset.public_access? ? 'Público' : 'Privado'
             %td.center= dataset.publish_date.strftime('%F') if dataset.publish_date
             %td.col-md-3