… Gemfile.lock
… app/views/inventories/shared/_distribution.html.haml