# Description

Reduces the response times from the backend. 

### Before
```
GET /instituciones/sedesol/catalogos
Completed 200 OK in 14416ms (Views: 11798.7ms | ActiveRecord: 622.1ms
```
```
GET /inventories/datasets
Completed 200 OK in 15514ms (Views: 15110.3ms | ActiveRecord: 386.4ms)
```

### After
```
GET /catalogos/548/conjuntos
Completed 200 OK in 2997ms (Views: 2826.4ms | ActiveRecord: 65.0ms)
```
```
GET /inventories/datasets
Completed 200 OK in 2629ms (Views: 2600.8ms | ActiveRecord: 9.3ms)
```